Split `Syscalls` into `Syscalls` and `RawSyscalls`.
`RawSyscalls` contains everything I previously called the "low-level API". `Syscalls` contains everything I previously called the "high-level API". A generic impl implements `Syscalls` for every type that implements `RawSyscalls`.
