+++ core/platform/src/lib.rs
@@ -10,10 +10,10 @@
 
 mod allows;
 mod error_code;
-mod syscall_types;
+mod raw_syscalls;
 mod syscalls;
 
 pub use allows::{AllowReadable, Allowed};
 pub use error_code::ErrorCode;
-pub use syscall_types::{OneArgMemop, ReturnType, YieldType, ZeroArgMemop};
+pub use raw_syscalls::{OneArgMemop, RawSyscalls, YieldType, ZeroArgMemop};
 pub use syscalls::Syscalls;
